Refactoring session_8 and small format on session_4 readme
