+++ Practices/session_8_Spark/practice_files/code/clinic_3.csv
