+++ Practices/session_8_Spark/practice_files/postgres/clinic.sql
