+++ Practices/session_8_Spark/pre-setup.md
@@ -1,0 +1,25 @@
+# Pre-setup
+
+## Prerequisites
+
+* [Docker][install_docker]
+* [DBeaver][dbeaver] \
+  **We will not be using the DBeaver during this document, however we will be using it during the session*
+
+## Steps
+
+### Step 1 - Composer
+
+* Start the Docker compose environment
+
+  ```sh
+  docker-compose up -d
+  ```
+
+## Links
+
+* [Install Docker][install_docker]
+* [DBeaver][dbeaver]
+
+[install_docker]: https://docs.docker.com/engine/install/
+[dbeaver]: https://dbeaver.io/download/
